+++ bean-searcher/src/main/java/com/ejlchina/searcher/implement/DefaultMetaResolver.java
@@ -2,11 +2,12 @@ package com.ejlchina.searcher.implement;
 
 import com.ejlchina.searcher.*;
 import com.ejlchina.searcher.bean.InheritType;
-import com.ejlchina.searcher.util.StringUtils;
 
 import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 
 /***
@@ -63,7 +64,11 @@ public class DefaultMetaResolver implements MetaResolver {
             if (column == null) {
                 continue;
             }
-            FieldMeta fieldMeta = resolveFieldMeta(beanMeta, column, field, index);
+            field.setAccessible(true);
+            SqlSnippet snippet = snippetResolver.resolve(column.getFieldSql());
+            // 注意：Oracle 数据库的别名不能以下划线开头
+            FieldMeta fieldMeta = new FieldMeta(beanMeta, field, snippet, "c_" + index,
+                    column.isConditional(), column.getOnlyOn());
             beanMeta.addFieldMeta(field.getName(), fieldMeta);
         }
         if (beanMeta.getFieldCount() == 0) {
@@ -87,22 +92,6 @@ public class DefaultMetaResolver implements MetaResolver {
             beanClass = beanClass.getSuperclass();
         }
         return fieldList.toArray(new Field[0]);
-    }
-
-    protected FieldMeta resolveFieldMeta(BeanMeta<?> beanMeta, DbMapping.Column column, Field field, int index) {
-        Method setter = getSetterMethod(beanMeta.getBeanClass(), field);
-        SqlSnippet snippet = snippetResolver.resolve(column.getFieldSql());
-        // 注意：Oracle 数据库的别名不能以下划线开头
-        return new FieldMeta(beanMeta, field, setter, snippet, "c_" + index, column.isConditional(), column.getOnlyOn());
-    }
-
-    protected Method getSetterMethod(Class<?> beanClass, Field field) {
-        String fieldName = field.getName();
-        try {
-            return beanClass.getMethod("set" + StringUtils.firstCharToUpperCase(fieldName), field.getType());
-        } catch (Exception e) {
-            throw new SearchException("[" + beanClass.getName() + ": " + fieldName + "] is annotated by @DbField, but there is no correctly setter for it.", e);
-        }
     }
 
     public SnippetResolver getSnippetResolver() {